fix(edgeless): sync selected rect
